refactor: change some html classes, imports and variables names
